# Conflicts:
#	src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/service/PipelineBuildService.kt
#	src/backend/ci/core/quality/biz-quality/src/main/kotlin/com/tencent/devops/quality/dao/v2/QualityControlPointDao.kt
#	src/backend/ci/core/quality/biz-quality/src/main/kotlin/com/tencent/devops/quality/dao/v2/QualityIndicatorDao.kt
#	src/backend/ci/core/quality/biz-quality/src/main/kotlin/com/tencent/devops/quality/resources/v2/UserQualityControlPointResourceImpl.kt
#	src/backend/ci/core/quality/biz-quality/src/main/kotlin/com/tencent/devops/quality/service/v2/QualityControlPointService.kt
#	src/backend/ci/core/quality/biz-quality/src/main/kotlin/com/tencent/devops/quality/service/v2/QualityIndicatorService.kt
#	src/backend/ci/core/quality/biz-quality/src/main/kotlin/com/tencent/devops/quality/service/v2/QualityMetadataService.kt
#	src/backend/ci/core/quality/biz-quality/src/main/kotlin/com/tencent/devops/quality/service/v2/QualityRuleCheckService.kt
#	src/backend/ci/core/quality/biz-quality/src/main/kotlin/com/tencent/devops/quality/service/v2/QualityTemplateService.kt

+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/interceptor/QueueInterceptor.kt
@@ -68,7 +68,7 @@ class QueueInterceptor @Autowired constructor(
                 // 设置了最大排队数量限制为0，但此时没有构建正在执行
                 Response(BuildStatus.RUNNING)
             } else if (maxQueue == 0 && buildSummaryRecord.runningCount > 0) {
-                Response(ERROR_PIPELINE_QUEUE_FULL, "流水线串行，排队数设置为0")
+                Response(ERROR_PIPELINE_QUEUE_FULL.toInt(), "流水线串行，排队数设置为0")
             } else if (buildSummaryRecord.queueCount >= maxQueue) {
                 // 排队数量超过最大限制
                 logger.info("[$pipelineId] MaxQueue=$maxQueue| currentQueue=${buildSummaryRecord.queueCount}")
